pyproject.toml
[project]
name = "synthesize-text2audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "accelerate",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "nltk>=3.9.1",
    "orjson>=3.11.2",
    "pydantic-settings>=2.10.1",
    "readability-lxml>=0.8.4.1",
    "scipy>=1.16.1",
    "torch==2.7.1",
    "transformers==4.52.4",
    "uvicorn[standard]>=0.35.0",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[dependency-groups]
dev = ["ruff>=0.12.9"]
